Move torch_trainer out of libmultilabel.
